added chrome stuff
trying to add a drop zone to drag and drop imgs to save to our website.
Also, added a context menu button if you right click an image and an
image only. add a variable for firebase linked to the one i set up and
added a section for firebase specific code. I’ll add you all as
developers to the firebase. I’ll be there in like an hour