Fully (partial) functioning extension
right-click saves to firebase. chrome exntesion button in browser shows
saved pictures (oldest to newest, may be able to change this, not
sure). Go to Mememaster page needs an absolute URL which we do not have
so for now takes you to google in a new tab. Other than that it works
great.